requirements.txt
# Core Python packages
#python_version >= "3.11", < "3.12"
numpy
pandas
# Google Gemini AI SDK
google-generativeai==0.8.5
google-genai==1.52.0
# Sentence Transformers for embeddings
sentence-transformers==5.1.2
streamlit==1.51.0
qdrant-client==1.16.1